+++ vkconfig_core/type_hide_message.cpp
@@ -42,6 +42,8 @@ const char* GetToken(HideMessageType value) {
                                 "HIDE_MESSAGE_WARN_CORE_SHADER_IGNORE",
                                 "HIDE_MESSAGE_WARN_OVERRIDE_LAYERS_SETTINGS",
                                 "HIDE_MESSAGE_WARN_OVERRIDE_LAYERS_SETTINGS_HIDE",
+                                "HIDE_MESSAGE_WARN_DARK_THEME_LINUX",
+                                "HIDE_MESSAGE_WARN_DARK_THEME_ICON",
                                 "HIDE_MESSAGE_QUESTION_REMOVING_LAYERS_PATH",
                                 "HIDE_MESSAGE_ERROR_32BIT"};
     static_assert(std::size(TOKENS) == HIDE_MESSAGE_COUNT);