+++ vkconfig_gui/tab_preferences.h
@@ -21,7 +21,6 @@
 #pragma once
 
 #include "tab.h"
-#include "../vkconfig_core/type_theme_mode.h"
 
 #include <QNetworkAccessManager>
 #include <QNetworkReply>
@@ -51,10 +50,9 @@ class TabPreferences : public Tab {
     void on_package_downloaded(QNetworkReply *pReply);
     void on_download_progress(qint64 bytesReceived, qint64 bytesTotal);
     void on_notify_releases_toggled(bool checked);
-    void on_theme_mode_changed();
+    void on_theme_mode_changed(int index);
 
    private:
     QNetworkAccessManager network_manager;
     QByteArray downloaded_data;
-    ThemeMode current_theme_mode = THEME_MODE_NONE;
 };